Renamed ChessServer\Command\Binary\Async as ChessServer\Command\Binary\Sync
