+++ src/Command/Binary/Cli.php
@@ -3,7 +3,7 @@
 namespace ChessServer\Command\Binary;
 
 use ChessServer\Command\AbstractCli;
-use ChessServer\Command\Binary\Async\ImageCommand;
+use ChessServer\Command\Binary\Sync\ImageCommand;
 use Spatie\Async\Pool;
 
 class Cli extends AbstractCli
